WebSocketModule (re-submit): fix secure websocket creation when protocols is null
Summary:
- [x] Explain the **motivation** for making this change.
- [x] Provide a **test plan** demonstrating that the code is solid.

## Motivation (required)

This PR fixes https://github.com/facebook/react-vr/issues/118. As Firebase uses `wss`, when combining firebase and react-vr shows that secure WebSockets are create with a `null` protocol parameter (https://developer.mozilla.org/en-US/docs/Web/API/WebSocket). As a consequence, the header `Sec-WebSocket-Protocol` is set to `null` which is not handled by firebase and make it responds with a 500 error code.

## Test Plan (required)

With a sample react-vr app, (https://github.com/alexduros/react-vr-app), I was able to test it with fixed `WebSocketModule` and the connexion with firebase is then ok.
Closes https://github.com/facebook/react-vr/pull/171

Differential Revision: D4976454

Pulled By: andrewimm

fbshipit-source-id: 64c1636