Push pixelRatio from client to Player
Summary:
Allow the pixelRatio to be set at in the client.js on the VRInstance.

This allows users to have better control of the fillrate aspects of the display. The default setting is to use the device pixel ratio.

Having a lower value will lower quality but improve performance for fill rate limitted applications

Reviewed By: lustyj

Differential Revision: D5059150

fbshipit-source-id: b8cff82