Route all mesh texture references through TextureManager
Summary: The `texture=` prop on all meshes (`Model` + primitives) now goes through TextureManager, allowing it to be cached centrally. This also means that custom textures can be directly referenced with the `texture()` helper. I've included an example that demonstrates how to implement custom textures, by rendering a fps meter in the runtime and attaching it onto a Box in React.

Reviewed By: mikearmstrong001

Differential Revision: D5060372

fbshipit-source-id: f6a50e5