Invalidate score if squares hidden. Document middlware limitations.
Summary:
clientPredictionConsistency only results in eventually
consistent clients if every ordering of a sequence of actions
which can cause a conflict does cause a conflict. This change
modifies the pairs validator to ensure that if a hide is
invalidated by a score action, then the score action is
invalidated by the hide to make this the case. It also adds
comments to document the restriction.

Reviewed By: gibsnson

Differential Revision: D5381180

fbshipit-source-id: 814bb98