Fix Oculus Browser link
Summary:
Point to the Oculus Browser store page. Also fixed up some minor formatting
as I was looking at fixing the VR Browsers page, but it isn't yet ready to
be fixed. The page mostly needs a rewrite to describe how to use VR browsers
now that the Carmel Developer Preview is not the way nor is the ovrweb
protocol.

Reviewed By: lustyj

Differential Revision: D5571363

fbshipit-source-id: 555787f