view-in-vr overlay
Summary:
Porting some of the OVRUI Overlay rendering logic over to this new layer.
Adds a listener to the VRState machine, to run callbacks whenever the current display changes. This allows rendering the View-In-VR button when a display is connected.
Moves the renderer canvas into a secondary wrapper, so that raycaster & input channel events can attach to that layer without interfering with the overlay.

Reviewed By: mikearmstrong001

Differential Revision: D7165834

fbshipit-source-id: 1cd6fe032915b7c524e50143adcf63b2f5d5639c