Register custom native modules with the new Runtime
Summary:
One of the challenges we encountered with the previous runtime was that some native modules depend on the ReactNativeContext for triggering any callbacks, but they also need to be initialized before the context. We solved this in the past by instrumenting these modules to have the context passed to them after they'd been registered. While adding native modules to the new runtime, I've addressed this in a cleaner way.

When declaring custom native modules, you can pass them the way previously supported, or you can pass a "Module Initializer." This is a simple method that takes a ReactNativeContext as its only argument, and returns a native module:

```
nativeModules: [
  (ctx) => new ModuleThatNeedsContext(ctx),
],
```

These get initialized within the Runtime, after the context has been created, but before it is initialized. This maintains the ideal behavior that we initialize the context as early as possible, to offset bundle load time.

Reviewed By: mikearmstrong001

Differential Revision: D7388693

fbshipit-source-id: 99f381bfadbcc65808a4588dbd705905f951b4a9