Expose static asset dir to client.js
Summary: While testing publishing apps, it occurred to me that users of the default starter project need to rewrite the background path as well as the asset root if they change the static assets location. To avoid this, I'd like to rewrite the starter project to use this new method, which references the assetRoot set at init time. That allows there to be one point of configuration for all properly-referenced assets on both the runtime and the React sides.

Reviewed By: mikearmstrong001

Differential Revision: D7822165

fbshipit-source-id: c16a55f5a82df4440fcbc893c4eb050fa86a280f